+++ src/functions/index.ts
@@ -1,5 +1,7 @@
 import { ModuleRegistrationContext } from '@causa/workspace';
+import { EmulatorStartForFirestore } from './emulator-start-firestore.js';
 import { EmulatorStartForPubSub } from './emulator-start-pubsub.js';
+import { EmulatorStopForFirestore } from './emulator-stop-firestore.js';
 import { EmulatorStopForPubSub } from './emulator-stop-pubsub.js';
 import { GoogleFirebaseStorageMergeRules } from './google-firebase-storage-merge-rules.js';
 import { GoogleFirestoreMergeRules } from './google-firestore-merge-rules.js';
@@ -10,7 +12,9 @@ import { SecretFetchForGoogleSecretManager } from './secret-fetch-secret-manager
 
 export function registerFunctions(context: ModuleRegistrationContext) {
   context.registerFunctionImplementations(
+    EmulatorStartForFirestore,
     EmulatorStartForPubSub,
+    EmulatorStopForFirestore,
     EmulatorStopForPubSub,
     GoogleFirebaseStorageMergeRules,
     GoogleFirestoreMergeRules,